add pause and resume
Allows to stop polling for events on a file descriptor until resumed
again. For example, if you have a source and a sink, but the sink is too
slow to keep up with data from the source buffer. This allows to pause the
source, until the sink catches up consuming the provided data.

Signed-off-by: Armin Preiml <apreiml@strohwolke.at>